refactor: split rule compression algos in a separate module
