+++ scripts/model-analyzer.js
@@ -3,8 +3,13 @@ const glob = require('glob')
 const path = require('path')
 const fs = require('fs')
 const yaml = require('yaml')
-const dGraph = require('dependency-graph')
 const Engine = require('publicodes').default
+const {
+	dGraphToJSON,
+	getCompressibleNodes,
+	getDependencyGraphStats,
+	getDependencyGraph,
+} = require('./rulesCompression')
 
 const line = () => console.log('---------------------------------------')
 
@@ -40,113 +45,6 @@ const args = yargs
 	.help()
 	.alias('help', 'h').argv
 
-console.log('args.model:', args.model)
-
-// Remove references to:
-// - $SITUATION rules
-// - parent namespaces
-const removeParentReferences = (dottedName, set) => {
-	const splittedDottedName = dottedName.split(' . ')
-
-	splittedDottedName
-		.slice(1, splittedDottedName.length)
-		.reduce((accName, name) => {
-			const newAccName = accName + ' . ' + name
-			set.delete(accName)
-			return newAccName
-		}, splittedDottedName[0])
-
-	set.delete(dottedName + ' . $SITUATION')
-
-	return set
-}
-
-const getDependencyGraphStats = (graph) => {
-	const graphEntries = Object.entries(graph.nodes)
-
-	const stats = graphEntries.reduce(
-		(stats, [name, _]) => {
-			const degIn = graph.directDependantsOf(name).length
-			const degOut = graph.directDependenciesOf(name).length
-			return {
-				...stats,
-				totalDeg: stats.totalDeg + degOut,
-				maxDegOut: degOut > stats.maxDegOut ? degOut : stats.maxDegOut,
-				maxDegIn: degIn > stats.maxDegIn ? degIn : stats.maxDegIn,
-			}
-		},
-		{
-			nbNodes: graph.size(),
-			totalDeg: 0,
-			averageDeg: 0,
-			maxDegOut: 0,
-			maxDegIn: 0,
-		}
-	)
-
-	return { ...stats, averageDeg: stats.totalDeg / stats.nbNodes }
-}
-
-const getDependencyGraph = (model, dependencyMap) => {
-	const graph = new dGraph.DepGraph()
-	var deps = Object.fromEntries(dependencyMap)
-	const modelWithoutSituations = Object.entries(model).filter(
-		([name, _]) => !name.endsWith('$SITUATION')
-	)
-
-	modelWithoutSituations.forEach(([name, value]) => {
-		graph.addNode(name, value)
-	})
-
-	Object.entries(deps)
-		.filter(([name, _]) => {
-			return !name.endsWith('$SITUATION')
-		})
-		.forEach(([name, set]) => {
-			const realDeps = removeParentReferences(name, set)
-			realDeps.forEach((to) => graph.addDependency(name, to))
-		})
-
-	return graph
-}
-
-const dGraphToJSON = (graph, ruleNames, parsedRules) => {
-	return ruleNames.reduce((acc, name) => {
-		const deps = graph.directDependantsOf(name)
-		const title = parsedRules[name]?.title ?? name
-
-		acc.push({ _name: name, _deps: deps, Title: title })
-		return acc
-	}, [])
-}
-
-// To be compressible, a node (rule) needs to be a constant:
-// - no question
-// - no dependency
-const isCompressible = (name, graph, parsedRules) => {
-	return (
-		!parsedRules[name].rawNode.question &&
-		graph.dependenciesOf(name).length === 0
-	)
-}
-
-const getCompressibleNodes = (graph, parsedRules) => {
-	var compressibleNodes = []
-	let compressibleLeafs = []
-
-	do {
-		compressibleLeafs = graph
-			.overallOrder(true)
-			.filter((leaf) => isCompressible(leaf, graph, parsedRules))
-		compressibleLeafs.forEach((leaf) => {
-			graph.removeNode(leaf)
-		})
-		compressibleNodes = compressibleNodes.concat(compressibleLeafs)
-	} while (compressibleLeafs.length > 0)
-
-	return { compressedGraph: graph, compressibleNodes }
-}
-
 glob(args.model, { ignore: args.ignore }, (err, files) => {
 	if (err) {
 		console.error(err)
@@ -163,10 +61,6 @@ glob(args.model, { ignore: args.ignore }, (err, files) => {
 		}
 	}, {})
 
-	line()
-	// console.log('Analyzed files:', files)
-	//
-	// line()
 	const ruleNames = Object.keys(model)
 	const engine = new Engine(model)
 	const parsedRules = engine.getParsedRules()
@@ -175,12 +69,10 @@ glob(args.model, { ignore: args.ignore }, (err, files) => {
 		engine.baseContext.referencesMaps.referencesIn
 	)
 	const beforeStats = getDependencyGraphStats(graph)
-	const { compressedGraph, compressibleNodes } = getCompressibleNodes(
-		graph,
-		parsedRules
-	)
+	const { compressedGraph } = getCompressibleNodes(graph, parsedRules)
 	const afterStats = getDependencyGraphStats(compressedGraph)
 
+	line()
 	console.log('Before compression:', beforeStats)
 	console.log('After compression:', afterStats)
